requirements.txt
Adafruit-DHT==1.3.2
RPi.GPIO==0.6.3
